fix moduleorder location
