+++ comp/system-probe/types/module.go
@@ -12,6 +12,9 @@ import (
 
 	"github.com/gorilla/mux"
 	"go.uber.org/fx"
+
+	"github.com/DataDog/datadog-agent/pkg/system-probe/config"
+	"github.com/DataDog/datadog-agent/pkg/system-probe/config/types"
 )
 
 // ErrNotEnabled is a special error type that should be returned by a Factory
@@ -33,4 +36,24 @@ type ProvidesSystemProbeModule struct {
 	fx.Out
 
 	Component SystemProbeModuleComponent `group:"systemprobe_module"`
+}
+
+// ModuleOrder is the desired creation order for system-probe modules
+var ModuleOrder = []types.ModuleName{
+	config.EBPFModule,
+	config.NetworkTracerModule,
+	config.TCPQueueLengthTracerModule,
+	config.OOMKillProbeModule,
+	config.EventMonitorModule, // there is a dependency from EventMonitor -> NetworkTracer, so EventMonitor has to follow NetworkTracer
+	config.ProcessModule,
+	config.DynamicInstrumentationModule, // dynamic instrumentation needs to be after EventMonitor
+	config.LanguageDetectionModule,
+	config.ComplianceModule,
+	config.PingModule,
+	config.TracerouteModule,
+	config.DiscoveryModule,
+	config.GPUMonitoringModule, // GPU monitoring needs to be initialized after EventMonitor, so that we have the event consumer ready
+	config.SoftwareInventoryModule,
+	config.PrivilegedLogsModule,
+	config.WindowsCrashDetectModule,
 }
